Adapt and overhaul SHP for new Component API
